+++ docs/conf.py
@@ -195,6 +195,6 @@ intersphinx_mapping = {'https://docs.python.org/': None}
 todo_include_todos = True
 
 autodoc_mock_imports = ["sip", "PyQt5", "PyQt5.sip", "PyQt5.QtCore", "PyQt5.QtGui", "PyQt5.QtCore", "qimage2ndarray",
-                        "pygame"]
+                        "pygame", "apng"]
 
 locale_dirs = ['locales/']