+++ easygraphics/processing/__init__.py
@@ -1,0 +1,5 @@
+from .processing import *
+from .processingwidget import *
+
+__all__ = processing.__all__
+__all__.append('ProcessingWidget')